requirements.txt
pymongo~=4.7.2
pydantic~=2.7.2
python-dotenv~=1.0.1